Split out a lot of functionality into features, still some warnings
