+++ crates/samples/echomain-stateful/Cargo.toml
@@ -27,3 +27,6 @@ features = [
 
 [dependencies.mssf-core]
 path = "../../libs/core"
+# We don't showcase config integration in this particular example
+default-features = false
+#features = []
